==> 2_analysis/5_NTRK/3_NTRK_fusion_prognosis.Rmd <==
---
title: "NTRK_fusion_prognosis"
author: "Yuki Saito"
date: "2025-01-04"
output: html_document
---

- R markdown script to evaluate prognosis by NTRK fusion status.

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r library, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readxl)
library(lubridate)
library(cowplot)
library(scales)
theme_set(theme_cowplot())
```

```{r source}
source("../../R/utils.R")
source("../../R/figure.R")
```

```{r make directory}
# make directory
dir.create("result/table/analysis/NTRK/", showWarnings = FALSE, recursive = TRUE)
dir.create("result/pdf/analysis/NTRK/", showWarnings = FALSE, recursive = TRUE)
```

```{r read data}
# read in merged data
merged_data <- read_csv("result/table/merged/merged_patient_data10.csv", na = "", guess_max = 55000) 
merged_data

NTRK_summary_df <- read_csv("result/table/analysis/NTRK/NTRK_fusion_cancer_type_count.csv", na = "")  
NTRK_summary_df
```

```{r check}
merged_data %>% 
  filter(str_detect(Evidence_A, "NTRK")) %>%
  count(Highest_evidence_level2)

merged_data %>% 
  filter(str_detect(Evidence_A, "NTRK")) %>%
  arrange(Registration_date)
```

```{r preparation}
merged_data2 <- merged_data %>% 
  filter(Pre_regimen_info_existence == "Yes") %>% 
  filter(is.na(PreEP_TRK)) %>% 
  filter(Oncotree_CODE_Final_wide %in% NTRK_summary_df$Oncotree_CODE_Final_wide) %>% 
  mutate(Age_category = case_when(Age < 65 ~ "<65",
                                  Age >= 65 & Age < 75 ~ "65-74",
                                  Age >= 75 ~ ">=75")) %>%
  mutate(Age_category = factor(Age_category, levels = c("<65", "65-74", ">=75"))) %>% 
  mutate(Sex_name = if_else(Sex == 9, NA_character_, Sex_name))

merged_data2 <- merged_data2 %>%
  mutate(class = case_when(str_detect(Evidence_A, "ETV6-NTRK3") ~ "ETV6-NTRK3", 
                            str_detect(Evidence_A, "NTRK") ~ "NTRK (other than ETV6-NTRK3)",
                            TRUE ~ "Others")) %>% 
  mutate(class = factor(class,
                         levels = c("ETV6-NTRK3", "NTRK (other than ETV6-NTRK3)", "Others")))
```

```{r survival}
library(ggsurvfit)
library(gtsummary)
library(survival)
library(survminer)

survdiff(Surv(OS_time_from_CGP, OS)~ class, data = merged_data2)
summary(coxph(Surv(OS_time_from_CGP, OS)~ class + Sex_name + Age_category, data = merged_data2)) 

ggsurvfit(survfit2(Surv(OS_time_from_CGP, OS) ~ class, data = merged_data2), size = 1.5) +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(xlim = c(0, 2),
                  ylim = c(0, 1)) +
  labs(x = "Years",
       y = "Overall Survival") +
  cowplot::theme_cowplot() +
  scale_color_manual(
    name = "Highest evidence level",
    values = c(
      "ETV6-NTRK3" = "#f8766d",
      "NTRK (other than ETV6-NTRK3)" = "#00ba38",
      "Others" = "gray60"
    )) +
  ggtitle("PanCan") 
ggsave("result/pdf/analysis/NTRK/NTRK_fusion_PanCan_prognosis.pdf", create.dir = TRUE)

merged_data2 <- merged_data2 %>%
  mutate(class = factor(class,
                         levels = c("NTRK (other than ETV6-NTRK3)", "ETV6-NTRK3", "Others")))
summary(coxph(Surv(OS_time_from_CGP, OS)~ class + Sex_name + Age_category, data = merged_data2))
summary(coxph(Surv(OS_time_from_CGP, OS)~ class + Sex_name + Age_category, data = merged_data2))$coefficients
```

```{r sessioninfo}
sessionInfo()
```
